Time: 4 ms (47.38%), Space: 97.1 MB (26.39%) - LeetHub

+++ 2358-number-of-ways-to-split-array/2358-number-of-ways-to-split-array.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int waysToSplitArray(vector<int>& nums) {
         int n = nums.size();
-        vector<long long> prefsum(n);
+        vector<long> prefsum(n);
         prefsum[0] += nums[0];
         for (int i{1}; i < n; i++)
             prefsum[i] += prefsum[i - 1] + nums[i];
